Moved IndexerFactories from xml to generic package
