+++ georocket-server/src/main/java/io/georocket/index/generic/GenericAttributeIndexerFactory.java
@@ -1,4 +1,4 @@
-package io.georocket.index.xml;
+package io.georocket.index.generic;
 
 import java.util.Map;
